+++ GHIElectronics.TinyCLR.Devices/Uart.cs
@@ -28,10 +28,10 @@ namespace GHIElectronics.TinyCLR.Devices.Uart {
         public void Flush() => this.Provider.Flush();
 
         public int Read(byte[] buffer) => this.Read(buffer, 0, buffer.Length);
-        public int Read(byte[] buffer, int offset, int length) => this.Provider.Read(buffer, (int)offset, (int)length);
+        public int Read(byte[] buffer, int offset, int length) => this.Provider.Read(buffer, offset, length);
 
         public int Write(byte[] buffer) => this.Read(buffer, 0, buffer.Length);
-        public int Write(byte[] buffer, int offset, int length) => this.Provider.Write(buffer, (int)offset, (int)length);
+        public int Write(byte[] buffer, int offset, int length) => this.Provider.Write(buffer, offset, length);
 
         public void ClearWriteBuffer() => this.Provider.ClearWriteBuffer();
         public void ClearReadBuffer() => this.Provider.ClearReadBuffer();
